// src/api/models/templateSectionResDTOSectionKind.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */

export type TemplateSectionResDTOSectionKind = typeof TemplateSectionResDTOSectionKind[keyof typeof TemplateSectionResDTOSectionKind];


export const TemplateSectionResDTOSectionKind = {
  DETAIL: 'DETAIL',
  ACHIEVEMENT: 'ACHIEVEMENT',
  TASK: 'TASK',
  PROBLEM_SOLVING: 'PROBLEM_SOLVING',
  LEARNING: 'LEARNING',
} as const;