// src/api/models/templateCatalogResDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { TemplateSectionResDTO } from './templateSectionResDTO';

export interface TemplateCatalogResDTO {
  sections: TemplateSectionResDTO[];
}