// src/api/models/templateSectionResDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { TemplateSectionResDTOAnchorSlotId } from './templateSectionResDTOAnchorSlotId';
import type { TemplateSectionResDTOSectionKind } from './templateSectionResDTOSectionKind';
import type { TemplateSlotResDTO } from './templateSlotResDTO';
import type { TemplateSubTemplateResDTO } from './templateSubTemplateResDTO';

export interface TemplateSectionResDTO {
  sectionKind: TemplateSectionResDTOSectionKind;
  /** @nullable */
  anchorSlotId?: TemplateSectionResDTOAnchorSlotId;
  /** level 4 카테고리 슬롯 */
  slots: TemplateSlotResDTO[];
  /** level 5 하위 템플릿 (없으면 빈 배열) */
  subTemplates: TemplateSubTemplateResDTO[];
}